Implemented DH_VehicleRepository with all CRUD operations

+++ DH_VehicleInventory.Infrastructure/Repositories/DH_VehicleRepository.cs
@@ -2,45 +2,60 @@
 using DH_VehicleInventory.Domain.VehicleAggregate.Entities;
 using DH_VehicleInventory.Infrastructure.Data;
 using Microsoft.EntityFrameworkCore;
+using System.Collections.Generic;
+using System.Threading.Tasks;
 
 namespace DH_VehicleInventory.Infrastructure.Repositories
 {
-    public class VehicleRepository : IVehicleRepository
+    public class DH_VehicleRepository : IVehicleRepository
     {
         private readonly DH_InventoryDbContext _context;
 
-        public VehicleRepository(DH_InventoryDbContext context)
+        public DH_VehicleRepository(DH_InventoryDbContext context)
         {
             _context = context ?? throw new ArgumentNullException(nameof(context));
         }
 
         public Vehicle Add(Vehicle vehicle)
         {
-            return _context.DH_Vehicles.Add(vehicle).Entity;
+            if (vehicle == null)
+                throw new ArgumentNullException(nameof(vehicle));
+
+            _context.DH_Vehicles.Add(vehicle);
+            _context.SaveChanges();
+
+            return vehicle;
         }
 
         public void Update(Vehicle vehicle)
         {
+            if (vehicle == null)
+                throw new ArgumentNullException(nameof(vehicle));
+
             _context.DH_Vehicles.Update(vehicle);
+            _context.SaveChanges();
         }
 
         public void Delete(Vehicle vehicle)
         {
+            if (vehicle == null)
+                throw new ArgumentNullException(nameof(vehicle));
+
             _context.DH_Vehicles.Remove(vehicle);
+            _context.SaveChanges();
         }
 
         public async Task<Vehicle> GetByIdAsync(int id)
         {
-            return await _context.DH_Vehicles
-                .Include(v => v.Inventories)
-                .FirstOrDefaultAsync(v => v.Id == id);
+            if (id <= 0)
+                throw new ArgumentException("Vehicle ID must be positive", nameof(id));
+
+            return await _context.DH_Vehicles.FindAsync(id);
         }
 
         public async Task<IEnumerable<Vehicle>> GetAllAsync()
         {
-            return await _context.DH_Vehicles
-                .Include(v => v.Inventories)
-                .ToListAsync();
+            return await _context.DH_Vehicles.ToListAsync();
         }
     }
 }
